+++ src/ll/orasql/scripts/oradrop.py
@@ -57,10 +57,19 @@ Options
 
 	``-e``, ``--encoding`` : encoding
 		The encoding of the output (if ``-x`` is not given; default is ``utf-8``).
+
+	``--include`` : regexp
+		Only include objects in the output if their name contains the regular
+		expression.
+
+	``--exclude`` : regexp
+		Exclude objects from the output if their name contains the regular
+		expression.
+
 """
 
 
-import sys, os, argparse
+import sys, os, re, argparse
 
 from ll import misc, astyle, orasql
 
@@ -84,6 +93,8 @@ def main(args=None):
 	p.add_argument("-k", "--keepjunk", dest="keepjunk", help="Output objects with '$' in their name? (default %(default)s)", action=misc.FlagAction, default=False)
 	p.add_argument("-i", "--ignore", dest="ignore", help="Ignore errors? (default %(default)s)", default=False, action=misc.FlagAction)
 	p.add_argument("-e", "--encoding", dest="encoding", help="Encoding for output (default %(default)s)", default="utf-8")
+	p.add_argument(      "--include", dest="include", metavar="REGEXP", help="Include only objects whose name contains PATTERN (default: %(default)s)", type=re.compile)
+	p.add_argument(      "--exclude", dest="exclude", metavar="REGEXP", help="Exclude objects whose name contains PATTERN (default: %(default)s)", type=re.compile)
 
 	args = p.parse_args(args)
 
@@ -106,9 +117,11 @@ def main(args=None):
 	def keep(obj):
 		if obj.owner is not None and not isinstance(obj, orasql.ForeignKey):
 			return False
-		if args.keepjunk:
-			return True
-		if "$" in obj.name:
+		if ("$" in obj.name or obj.name.startswith("SYS_EXPORT_SCHEMA_")) and not args.keepjunk:
+			return False
+		if args.include is not None and args.include.search(obj.name) is None:
+			return False
+		if args.exclude is not None and args.exclude.search(obj.name) is not None:
 			return False
 		return True