+++ src/ll/orasql/scripts/oragrant.py
@@ -57,6 +57,14 @@ Options
 	``-e``, ``--encoding`` : encoding
 		The encoding of the output (if ``-x`` is not given; default is ``utf-8``).
 
+	``--include`` : regexp
+		Only include objects in the output if their name contains the regular
+		expression.
+
+	``--exclude`` : regexp
+		Exclude objects from the output if their name contains the regular
+		expression.
+
 
 Example
 -------
@@ -68,7 +76,7 @@ Grant all privileges that ``alice`` has in the schema ``user@db`` to ``bob`` in
 """
 
 
-import sys, os, argparse
+import sys, os, re, argparse
 
 from ll import misc, astyle, orasql
 
@@ -92,6 +100,8 @@ def main(args=None):
 	p.add_argument("-i", "--ignore", dest="ignore", help="Ignore errors? (default %(default)s)", default=False, action=misc.FlagAction)
 	p.add_argument("-m", "--mapgrantee", dest="mapgrantee", help="Map grantees (Python expression: list or dict)", default="True")
 	p.add_argument("-e", "--encoding", dest="encoding", help="Encoding for output (default %(default)s)", default="utf-8")
+	p.add_argument(      "--include", dest="include", metavar="REGEXP", help="Include only objects whose name contains PATTERN (default: %(default)s)", type=re.compile)
+	p.add_argument(      "--exclude", dest="exclude", metavar="REGEXP", help="Exclude objects whose name contains PATTERN (default: %(default)s)", type=re.compile)
 
 	args = p.parse_args(args)
 
@@ -120,9 +130,11 @@ def main(args=None):
 	mapgrantee = eval(args.mapgrantee)
 
 	def keep(obj):
-		if args.keepjunk:
-			return True
-		if "$" in obj.name or "/" in obj.name or obj.name.startswith("SYS_EXPORT_SCHEMA_"):
+		if ("$" in obj.name or "/" in obj.name or obj.name.startswith("SYS_EXPORT_SCHEMA_")) and not args.keepjunk:
+			return False
+		if args.include is not None and args.include.search(obj.name) is None:
+			return False
+		if args.exclude is not None and args.exclude.search(obj.name) is not None:
 			return False
 		return True